Refactor bin/commands
