+++ bin/commands
@@ -87,95 +87,73 @@ display_name () {
     fi
 }
 
+display_detail_section () {
+    if [ -n "${2}" ]; then
+        echo -n "${INDENT1}"
+        print_yellow "${1}:"
+        echo
+        echo "${INDENT2}${2//${LF}/${LF}${INDENT2}}"
+        echo
+    fi
+}
+
 display_description () {
     if [ "$1" = "-n" ]; then
         if [ -n "$2" ]; then
-            echo " : ${2//${LF}*/}"
+            DESCRIPTION=$(get_description ${2})
+            echo " : ${DESCRIPTION//${LF}*/}"
         else
             echo
         fi
     else
-        if [ -n "${1}" ]; then
-            echo -n "${INDENT1}"
-            print_yellow "Description:"
-            echo
-            echo "${INDENT2}${1//${LF}/${LF}${INDENT2}}"
-            echo
-        fi
+        DESCRIPTION=$(get_description ${1})
+        display_detail_section "Description" "${DESCRIPTION}"
     fi
 }
 
 display_usage () {
-    if [ -n "${1}" ]; then
-        echo -n "${INDENT1}"
-        print_yellow "Usage:"
-        echo
-        echo "${INDENT2}${1//${LF}/${LF}${INDENT2}}"
-        echo
-    fi
+    USAGE=$(get_usage ${1})
+    display_detail_section "Usage" "${USAGE}"
 }
 
 display_arguments () {
-    if [ -n "${1}" ]; then
-        echo -n "${INDENT1}"
-        print_yellow "Arguments:"
-        echo
-        echo "${INDENT2}${1//${LF}/${LF}${INDENT2}}"
-        echo
-    fi
+    ARGUMENTS=$(get_arguments ${1})
+    display_detail_section "Arguments" "${ARGUMENTS}"
 }
 
 display_options () {
-    if [ -n "${1}" ]; then
-        echo -n "${INDENT1}"
-        print_yellow "Options:"
-        echo
-        echo "${INDENT2}${1//${LF}/${LF}${INDENT2}}"
-        echo
-    fi
+    OPTIONS=$(get_options ${1})
+    display_detail_section "Options" "${OPTIONS}"
 }
 
 display_example () {
-    if [ -n "${1}" ]; then
-        echo -n "${INDENT1}"
-        print_yellow "Example: "
-        echo
-        echo "${INDENT2}${1//${LF}/${LF}${INDENT2}}"
-        echo
-    fi
+    EXAMPLE=$(get_example ${1})
+    display_detail_section "Example" "${EXAMPLE}"
 }
 
 display_list () {
     if [ -n "${TARGET}" ] && [[ "${1}" != "${TARGET}"* ]]; then
         return
     fi
-    # attributes
-    DESCRIPTION=$(get_description ${1})
 
     # command name
     display_name -n "${1}"
 
     # description
-    display_description -n "${DESCRIPTION}"
+    display_description -n "${1}"
 }
 
 display_detail () {
     if [ -n "${TARGET}" ] && [[ "${1}" != "${TARGET}"* ]]; then
         return
     fi
-    # attributes
-    DESCRIPTION=$(get_description ${1})
-    USAGE=$(get_usage ${1})
-    ARGUMENTS=$(get_arguments ${1})
-    OPTIONS=$(get_options ${1})
-    EXAMPLE=$(get_example ${1})
 
     display_name "${1}"
-    display_description "${DESCRIPTION}"
-    display_usage "${USAGE}"
-    display_arguments "${ARGUMENTS}"
-    display_options "${OPTIONS}"
-    display_example "${EXAMPLE}"
+    display_description "${1}"
+    display_usage "${1}"
+    display_arguments "${1}"
+    display_options "${1}"
+    display_example "${1}"
 }
 
 get_args $@
