+++ docker/laravel/app/Console/Commands/MinioPutCommand.php
