+++ docker/laravel/app/Console/Commands/MinitGetCommand.php
