+++ docker/laravel/config/filesystems.php
